cleaned up restPOJO and rename Class. +strgalt l in files
