+++ src/main/java/com/application/data/restPOJO/Distance.java
@@ -2,9 +2,9 @@
 package com.application.data.restPOJO;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import javax.annotation.Generated;
+
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -14,73 +14,35 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-    "destination_addresses",
-    "origin_addresses",
-    "rows",
-    "status"
+        "text",
+        "value"
 })
 @Generated("jsonschema2pojo")
 public class Distance {
 
-    @JsonProperty("destination_addresses")
-    private List<String> destinationAddresses = null;
-    @JsonProperty("origin_addresses")
-    private List<String> originAddresses = null;
-    @JsonProperty("rows")
-    private List<Row> rows = null;
-    @JsonProperty("status")
-    private String status;
-    @JsonIgnore
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
+    @JsonProperty("text")
+    private String text;
+    @JsonProperty("value")
+    private Integer value;
 
-    @JsonProperty("destination_addresses")
-    public List<String> getDestinationAddresses() {
-        return destinationAddresses;
+    @JsonProperty("text")
+    public String getText() {
+        return text;
     }
 
-    @JsonProperty("destination_addresses")
-    public void setDestinationAddresses(List<String> destinationAddresses) {
-        this.destinationAddresses = destinationAddresses;
+    @JsonProperty("text")
+    public void setText(String text) {
+        this.text = text;
     }
 
-    @JsonProperty("origin_addresses")
-    public List<String> getOriginAddresses() {
-        return originAddresses;
+    @JsonProperty("value")
+    public Integer getValue() {
+        return value;
     }
 
-    @JsonProperty("origin_addresses")
-    public void setOriginAddresses(List<String> originAddresses) {
-        this.originAddresses = originAddresses;
-    }
-
-    @JsonProperty("rows")
-    public List<Row> getRows() {
-        return rows;
-    }
-
-    @JsonProperty("rows")
-    public void setRows(List<Row> rows) {
-        this.rows = rows;
-    }
-
-    @JsonProperty("status")
-    public String getStatus() {
-        return status;
-    }
-
-    @JsonProperty("status")
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
+    @JsonProperty("value")
+    public void setValue(Integer value) {
+        this.value = value;
     }
 
 }
