+++ src/main/java/com/application/data/restPOJO/GoogleDist.java
@@ -1,0 +1,75 @@
+
+package com.application.data.restPOJO;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import javax.annotation.Generated;
+
+import com.fasterxml.jackson.annotation.JsonAnyGetter;
+import com.fasterxml.jackson.annotation.JsonAnySetter;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonPropertyOrder({
+        "destination_addresses",
+        "origin_addresses",
+        "rows",
+        "status"
+})
+@Generated("jsonschema2pojo")
+public class GoogleDist {
+
+    @JsonProperty("destination_addresses")
+    private List<String> destinationAddresses = null;
+    @JsonProperty("origin_addresses")
+    private List<String> originAddresses = null;
+    @JsonProperty("rows")
+    private List<Row> rows = null;
+    @JsonProperty("status")
+    private String status;
+
+    @JsonProperty("destination_addresses")
+    public List<String> getDestinationAddresses() {
+        return destinationAddresses;
+    }
+
+    @JsonProperty("destination_addresses")
+    public void setDestinationAddresses(List<String> destinationAddresses) {
+        this.destinationAddresses = destinationAddresses;
+    }
+
+    @JsonProperty("origin_addresses")
+    public List<String> getOriginAddresses() {
+        return originAddresses;
+    }
+
+    @JsonProperty("origin_addresses")
+    public void setOriginAddresses(List<String> originAddresses) {
+        this.originAddresses = originAddresses;
+    }
+
+    @JsonProperty("rows")
+    public List<Row> getRows() {
+        return rows;
+    }
+
+    @JsonProperty("rows")
+    public void setRows(List<Row> rows) {
+        this.rows = rows;
+    }
+
+    @JsonProperty("status")
+    public String getStatus() {
+        return status;
+    }
+
+    @JsonProperty("status")
+    public void setStatus(String status) {
+        this.status = status;
+    }
+
+}
