+++ src/main/java/com/application/data/restPOJO/Units.java
@@ -4,6 +4,7 @@ package com.application.data.restPOJO;
 import java.util.HashMap;
 import java.util.Map;
 import javax.annotation.Generated;
+
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -13,16 +14,16 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-    "time",
-    "predictability",
-    "precipitation_probability",
-    "pressure",
-    "relativehumidity",
-    "co",
-    "temperature",
-    "winddirection",
-    "precipitation",
-    "windspeed"
+        "time",
+        "predictability",
+        "precipitation_probability",
+        "pressure",
+        "relativehumidity",
+        "co",
+        "temperature",
+        "winddirection",
+        "precipitation",
+        "windspeed"
 })
 @Generated("jsonschema2pojo")
 public class Units {
@@ -47,8 +48,6 @@ public class Units {
     private String precipitation;
     @JsonProperty("windspeed")
     private String windspeed;
-    @JsonIgnore
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
     @JsonProperty("time")
     public String getTime() {
@@ -148,16 +147,6 @@ public class Units {
     @JsonProperty("windspeed")
     public void setWindspeed(String windspeed) {
         this.windspeed = windspeed;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
     }
 
 }
