+++ src/main/java/com/application/data/restPOJO/Weather.java
@@ -4,6 +4,7 @@ package com.application.data.restPOJO;
 import java.util.HashMap;
 import java.util.Map;
 import javax.annotation.Generated;
+
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -13,9 +14,9 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-    "metadata",
-    "units",
-    "data_day"
+        "metadata",
+        "units",
+        "data_day"
 })
 @Generated("jsonschema2pojo")
 public class Weather {
