+++ src/main/java/com/application/data/service/WeatherService.java
@@ -8,31 +8,31 @@ import org.springframework.web.reactive.function.client.WebClient;
 @Service
 public class WeatherService {
     private final WebClient webClient;
-    private final String key  = "l6gRrOTmayCWKws3";
+    private final String key = "l6gRrOTmayCWKws3";
 
-    public WeatherService(WebClient.Builder builder){
+    public WeatherService(WebClient.Builder builder) {
         webClient = builder.baseUrl("http://my.meteoblue.com/packages/").build();
     }
 
-    public Weather getForecast(String latitude, String longitude){
+    public Weather getForecast(String latitude, String longitude) {
         return webClient.get()
                 .uri(uriBuilder -> uriBuilder
                         .path("basic-day")
                         .queryParam("lat", latitude)
-                        .queryParam("lon",longitude)
-                        .queryParam("apikey",key)
+                        .queryParam("lon", longitude)
+                        .queryParam("apikey", key)
                         .build())
                 .retrieve()
                 .bodyToMono(Weather.class).block();
     }
 
-    public String getForecastString(String latitude, String longitude){
+    public String getForecastString(String latitude, String longitude) {
         return webClient.get()
                 .uri(uriBuilder -> uriBuilder
                         .path("basic-day")
                         .queryParam("lat", latitude)
-                        .queryParam("lon",longitude)
-                        .queryParam("apikey",key)
+                        .queryParam("lon", longitude)
+                        .queryParam("apikey", key)
                         .build())
                 .retrieve()
                 .bodyToMono(String.class).block();
